# components/webui/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Server",
    "description": "API Server for CLP",
    "contact": {
      "name": "YScope"
    },
    "license": {
      "name": ""
    },
    "version": "0.13.1-dev"
  },
  "paths": {
    "/compression/jobs": {
      "post": {
        "tags": [],
        "description": "Submits a compression job.",
        "operationId": "compression_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompressionJobCreation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The created compression job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompressionJob"
                }
              }
            }
          },
          "400": {
            "description": "Invalid dataset name"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/metadata/column_metadata/{dataset_name}/timestamp": {
      "get": {
        "tags": [],
        "description": "Gets the timestamp column names for a given dataset (CLP-S only).",
        "operationId": "timestamp_column_names",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid dataset name"
          },
          "404": {
            "description": "Dataset not found"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/metadata/compression_jobs": {
      "get": {
        "tags": [],
        "description": "Gets recent compression-job metadata (most recent first), with the decoded CLP IO config for each job.",
        "operationId": "compression_metadata",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompressionMetadata"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/metadata/datasets": {
      "get": {
        "tags": [],
        "description": "Gets the names of all datasets.",
        "operationId": "datasets",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/metadata/ingestion_details": {
      "get": {
        "tags": [],
        "description": "Gets ingestion details (timestamp range, file count, message count) across all datasets. Returns `null` when no data has been ingested yet.",
        "operationId": "ingestion_details",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/IngestionDetails"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/metadata/query_speed": {
      "get": {
        "tags": [],
        "description": "Gets the query speed (total uncompressed bytes scanned and job duration) for a search job across the given datasets. Returns `null` when the job hasn't scanned any archives or hasn't finished yet.",
        "operationId": "query_speed",
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "description": "Comma-separated list of dataset names (CLP-S only). Ignored for the CLP storage engine.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search_job_id",
            "in": "query",
            "description": "The search job ID whose scan speed should be computed.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/QuerySpeed"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid dataset name"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/metadata/space_savings": {
      "get": {
        "tags": [],
        "description": "Gets aggregated space-savings statistics (total uncompressed and compressed sizes) across all datasets.",
        "operationId": "space_savings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceSavings"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/metadata/time_range": {
      "get": {
        "tags": [],
        "description": "Gets the earliest and latest log entry timestamps across the given datasets. For the CLP storage engine, the `dataset` parameter is ignored. Returns `null` when no archives exist yet.",
        "operationId": "time_range",
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "description": "Comma-separated list of dataset names (CLP-S only). Ignored for the CLP storage engine.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/TimeRange"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid dataset name"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/os/ls": {
      "get": {
        "tags": [],
        "description": "Lists files and directories at the specified path.",
        "operationId": "list_files",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "The absolute filesystem path to list.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DirEntry"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Path not found"
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/query": {
      "post": {
        "tags": [],
        "description": "Submits a new query job.",
        "operationId": "query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryConfig"
              },
              "example": {
                "query_string": "*",
                "datasets": [
                  "default"
                ],
                "time_range_begin_millisecs": 0,
                "time_range_end_millisecs": 17356896,
                "ignore_case": true,
                "max_num_results": 0,
                "buffer_results_in_mongodb": true,
                "count_by_time_bucket_size_millisecs": null
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The URI to fetch the results of the submitted query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResultsUri"
                },
                "example": {
                  "query_results_uri": "query_results/1"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompressionJob": {
        "type": "object",
        "description": "Response body containing the ID of a newly created compression job.",
        "required": [
          "job_id"
        ],
        "properties": {
          "job_id": {
            "type": "integer",
            "format": "int64",
            "description": "The ID of the newly created compression job."
          }
        }
      },
      "CompressionJobCreation": {
        "type": "object",
        "description": "Request body for submitting a compression job.",
        "required": [
          "paths"
        ],
        "properties": {
          "dataset": {
            "type": [
              "string",
              "null"
            ],
            "description": "Dataset to compress into (CLP-S only). Optional for the CLP storage engine."
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Absolute filesystem paths of the files to compress."
          },
          "timestamp_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Timestamp key to use when parsing logs."
          },
          "unstructured": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the input is unstructured. Defaults to `true` for CLP and `false` for CLP-S."
          }
        },
        "additionalProperties": false
      },
...
      "QueryResultsUri": {
        "type": "object",
        "required": [
          "query_results_uri"
        ],
        "properties": {
          "query_results_uri": {
            "type": "string",
            "description": "The uri to get the query results."
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}